+++ src/transformers/models/granite/__init__.py
@@ -1,3 +1,63 @@
-from .base import GraniteModel, GranitePreTrainedModel
-from .config import GraniteConfig
-from .main import GraniteForCausalLM
+# Copyright 2023 The HuggingFace Team. All rights reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+from typing import TYPE_CHECKING
+
+from ...utils import (
+    OptionalDependencyNotAvailable,
+    _LazyModule,
+    is_torch_available,
+)
+
+
+_import_structure = {
+    "configuration_granite": ["GraniteConfig"],
+}
+
+try:
+    if not is_torch_available():
+        raise OptionalDependencyNotAvailable()
+except OptionalDependencyNotAvailable:
+    pass
+else:
+    _import_structure["modeling_granite"] = [
+        "GraniteForCausalLM",
+        "GraniteForSequenceClassification",
+        "GraniteForTokenClassification",
+        "GraniteModel",
+        "GranitePreTrainedModel",
+    ]
+
+if TYPE_CHECKING:
+    from .configuration_granite import GraniteConfig
+
+    try:
+        if not is_torch_available():
+            raise OptionalDependencyNotAvailable()
+    except OptionalDependencyNotAvailable:
+        pass
+    else:
+        from .modeling_granite import (
+            GraniteForCausalLM,
+            GraniteForSequenceClassification,
+            GraniteForTokenClassification,
+            GraniteModel,
+            GranitePreTrainedModel,
+        )
+
+
+else:
+    import sys
+
+    sys.modules[__name__] = _LazyModule(__name__, globals()["__file__"], _import_structure, module_spec=__spec__)
