+++ src/transformers/models/granite/modeling_granite/__init__.py
@@ -10,16 +10,17 @@ import torch
 import torch.nn as nn
 from torch.utils.checkpoint import checkpoint
 
-from ...modeling_outputs import BaseModelOutputWithPastAndCrossAttentions
-from ...modeling_utils import PreTrainedModel
+from ....modeling_outputs import BaseModelOutputWithPastAndCrossAttentions
+from ....modeling_utils import PreTrainedModel
 from .attention import Attention
-from .config import GraniteConfig
+from ..configuration_granite import GraniteConfig
 from .enums import AttentionHeadType, AttentionImplementation, NormalizationImplementation, PositionEmbeddingType
 from .layer import GraniteBlock
 from .mlp import MLP
 from .normalization import RMSNorm, get_normalization_function
 from .position_embedding import Alibi, RoPE, YaRNScaledRoPE
 from .utils import check_list_type, flatten_and_convert_to_tensors
+from ....modeling_outputs import CausalLMOutputWithCrossAttentions
 
 
 class GranitePreTrainedModel(PreTrainedModel):
@@ -691,3 +692,160 @@ class GraniteModel(GranitePreTrainedModel):
                 )
         else:
             raise NotImplementedError()
+
+
+class GraniteForCausalLM(GranitePreTrainedModel):
+    _keys_to_ignore_on_load_missing = [r"attn.masked_bias", r"attn.bias", r"lm_head.weight"]
+
+    def __init__(self, config: GraniteConfig, **kwargs) -> None:
+        super().__init__(config, **kwargs)
+        self.transformer = GraniteModel(config, **kwargs)
+        self.lm_head = nn.Linear(config.n_embd, config.vocab_size, bias=False)
+
+        # Initialize weights and apply final processing
+        self.post_init()
+
+    def get_input_embeddings(self) -> nn.Embedding:
+        return self.transformer.wte
+
+    def set_input_embeddings(self, value: nn.Embedding) -> None:
+        self.transformer.wte = value
+
+    def get_output_embeddings(self) -> nn.Linear:
+        return self.lm_head
+
+    def set_output_embeddings(self, new_embeddings: nn.Linear) -> None:
+        self.lm_head = new_embeddings
+
+    # FIXME typing
+    def prepare_inputs_for_generation(
+        self, input_ids: torch.Tensor, past_key_values=None, inputs_embeds: torch.Tensor = None, **kwargs
+    ) -> dict:
+        token_type_ids = kwargs.get("token_type_ids", None)
+        # Omit tokens covered by past_key_values
+        if past_key_values:
+            past_length = past_key_values[0][0].shape[2]
+
+            # Some generation methods already pass only the last input ID
+            if input_ids.shape[1] > past_length:
+                remove_prefix_length = past_length
+            else:
+                # Default to old behavior: keep only final ID
+                remove_prefix_length = input_ids.shape[1] - 1
+
+            input_ids = input_ids[:, remove_prefix_length:]
+            if token_type_ids is not None:
+                token_type_ids = token_type_ids[:, -input_ids.shape[1] :]
+
+        attention_mask = kwargs.get("attention_mask", None)
+        position_ids = kwargs.get("position_ids", None)
+
+        if attention_mask is not None and position_ids is None:
+            # create position_ids on the fly for batch generation
+            position_ids = attention_mask.long().cumsum(-1) - 1
+            position_ids.masked_fill_(attention_mask == 0, 0)
+            if past_key_values:
+                position_ids = position_ids[:, -input_ids.shape[1] :]
+        else:
+            position_ids = None
+
+        # if `inputs_embeds` are passed, we only want to use them in the 1st generation step
+        if inputs_embeds is not None and past_key_values is None:
+            model_inputs = {"inputs_embeds": inputs_embeds}
+        else:
+            model_inputs = {"input_ids": input_ids}
+
+        model_inputs.update(
+            {
+                "past_key_values": past_key_values,
+                "use_cache": kwargs.get("use_cache"),
+                "position_ids": position_ids,
+                "attention_mask": attention_mask,
+                "token_type_ids": token_type_ids,
+            }
+        )
+        return model_inputs
+
+    def forward(
+        self,
+        input_ids: Union[torch.Tensor, List[List[int]]] = None,
+        past_key_values: Tuple[Tuple[torch.Tensor]] = None,
+        attention_mask: torch.Tensor = None,
+        token_type_ids: Union[torch.Tensor, List[List[int]]] = None,
+        position_ids: Union[torch.Tensor, List[List[int]]] = None,
+        head_mask: torch.Tensor = None,
+        inputs_embeds: Union[torch.Tensor, List[List[float]]] = None,
+        labels: Union[torch.Tensor, List[List[int]]] = None,
+        use_cache: bool = None,
+        output_attentions: bool = None,
+        output_hidden_states: bool = None,
+        return_dict: bool = None,
+        cu_seqlens: torch.Tensor = None,
+        max_seqlen: torch.Tensor = None,
+    ) -> Union[Tuple, CausalLMOutputWithCrossAttentions]:
+        return_dict = return_dict if return_dict is not None else self.config.use_return_dict
+
+        input_ids, position_ids, token_type_ids, labels, cu_seqlens, max_seqlen = self.prepare_inputs_for_model(
+            input_ids=input_ids,
+            inputs_embeds=inputs_embeds,
+            position_ids=position_ids,
+            token_type_ids=token_type_ids,
+            labels=labels,
+            cu_seqlens=cu_seqlens,
+            max_seqlen=max_seqlen,
+            past_key_values=past_key_values,
+            attention_mask=attention_mask,
+            head_mask=head_mask,
+            use_cache=use_cache,
+            output_attentions=output_attentions,
+            attention_implementation=self.attention_implementation,
+        )
+
+        # ==========================================================================================
+        # AttentionImplementation.padding_free:
+        #     input_ids -> (total_q)
+        #     attention_mask -> None
+        #     position_ids -> (total_q)
+        # else:
+        #     input_ids -> (batch_size, query_length)
+        #     attention_mask -> None or (batch_size, key_length)
+        #     position_ids -> None or (batch_size, key_length)
+        # ==========================================================================================
+
+        transformer_outputs = self.transformer(
+            input_ids,
+            past_key_values=past_key_values,
+            attention_mask=attention_mask,
+            token_type_ids=token_type_ids,
+            position_ids=position_ids,
+            head_mask=head_mask,
+            inputs_embeds=inputs_embeds,
+            use_cache=use_cache,
+            output_attentions=output_attentions,
+            output_hidden_states=output_hidden_states,
+            return_dict=return_dict,
+            cu_seqlens=cu_seqlens,
+            max_seqlen=max_seqlen,
+        )
+        hidden_states = transformer_outputs[0]
+
+        lm_logits = self.lm_head(hidden_states)
+        loss = self.get_autoregressive_language_modeling_loss(lm_logits, labels, cu_seqlens)
+
+        if not return_dict:
+            output = (lm_logits,) + transformer_outputs[1:]
+            return ((loss,) + output) if loss is not None else output
+
+        return CausalLMOutputWithCrossAttentions(
+            loss=loss,
+            logits=lm_logits,
+            past_key_values=transformer_outputs.past_key_values,
+            hidden_states=transformer_outputs.hidden_states,
+            attentions=transformer_outputs.attentions,
+        )
+
+    @staticmethod
+    def _reorder_cache(
+        past_key_values: Tuple[Tuple[torch.Tensor]], beam_idx: torch.Tensor
+    ) -> Tuple[Tuple[torch.Tensor]]:
+        return tuple(layer_past.index_select(0, beam_idx.to(layer_past.device)) for layer_past in past_key_values)
