+++ src/transformers/models/granite/modeling_granite/activations.py
@@ -1,7 +1,7 @@
 import torch
 import torch.nn as nn
 
-from ...activations import get_activation as get_base_activation
+from ....activations import get_activation as get_base_activation
 
 
 _GLU_BASE_MAPPING = {
