+++ src/transformers/models/granite/modeling_granite/attention/base.py
