+++ src/transformers/models/granite/modeling_granite/attention/flash.py
