+++ src/transformers/models/granite/modeling_granite/attention/math.py
