+++ src/transformers/models/granite/modeling_granite/mlp.py
