+++ src/transformers/models/granite/modeling_granite/normalization/layernorm/apex_persistent.py
