+++ src/transformers/models/granite/modeling_granite/normalization/rmsnorm/apex.py
