+++ src/transformers/models/granite/modeling_granite/normalization/rmsnorm/base.py
