+++ src/transformers/models/granite/modeling_granite/position_embedding/rope.py
