+++ src/transformers/models/granite/modeling_granite/utils.py
